refactored Analytics component
